test(exercise): add exercise API integration tests; fix TINYINT migrations
Integration tests cover full HTTP stack with Testcontainers:
- POST /api/v1/exercises: create, duplicate name 409, blank name 400, no auth 401
- GET /api/v1/exercises/{id}: found 200, not found 404
- PUT /api/v1/exercises/{id}: update 200, forbidden 403 (other user), not found 404
- DELETE /api/v1/exercises/{id}: delete 204, forbidden 403, verify deleted 404
- GET /api/v1/exercises: list 200, search filter 200
- GET /api/v1/exercises/categories: 200
- GET /api/v1/exercises/muscle-groups: 200 with all 10 groups

Fix pre-existing migration bug: TINYINT is not a valid PostgreSQL type.
Replace TINYINT with BIGINT in V6 (muscles), V9 (exercise_categories),
and V10 (exercises.category_id) to match JPA entity Long field mapping.

+++ src/main/resources/db/changelog/V6__create_muscles_table.sql
@@ -2,7 +2,7 @@
 
 --changeset liftit:create-muscles-table
 CREATE TABLE muscles (
-    id         TINYINT                  NOT NULL,
+    id         BIGINT                   NOT NULL,
     name       VARCHAR(20)              NOT NULL,
     created_at TIMESTAMP WITH TIME ZONE NOT NULL,
     created_by BIGINT                   NOT NULL,
